pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "heddle"
version = "0.1.0"
description = "Heddle — A secure, declarative MCP runtime"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "fastmcp>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "structlog>=24.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[project.scripts]
heddle = "heddle.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
